fix(loc-map): fixed localization and mapping include headers
