+++ src/loc_map/Doxyfile
