+++ src/loc_map/include/utils/color.hpp
@@ -1,0 +1,13 @@
+#ifndef SRC_LOC_MAP_INCLUDE_UTILS_COLOR_HPP_
+#define SRC_LOC_MAP_INCLUDE_UTILS_COLOR_HPP_
+
+/**
+ * @brief Enum for colors
+ *
+ */
+namespace colors {
+enum Color { red = 0, yellow, blue };
+static const char* color_names[] = {"red", "yellow", "blue"};
+}  // namespace colors
+
+#endif  // SRC_LOC_MAP_INCLUDE_UTILS_COLOR_HPP_
